Its done
i found a way to make the air control work great, the main problem was that once one part of the in air movement went above a certain limit the player could no longer add speed, which is the idea, but it also meant the player could not use any other controls to go back, to fix this i made it so if the player went above the limit on one of the axis the controls would minus speed instead of add, but that meant when the player regained control the speeds would be opposite or something.

anyway i made the InAirMovement half itself if it went too far over the limit and it works great. I will polish up the code later and add wall running when my brain isnt hurting.